Code refactoring for the top consumer and inventory filtering
