+++ frontend/packages/ceph-storage-plugin/src/components/dashboard-page/storage-dashboard/top-consumers-card/top-consumers-card-body.tsx
@@ -15,7 +15,7 @@ import { DataPoint, PrometheusResponse } from '@console/internal/components/grap
 import { humanizeBinaryBytesWithoutB, LoadingInline } from '@console/internal/components/utils';
 import { twentyFourHourTime } from '@console/internal/components/utils/datetime';
 import { GraphEmpty } from '@console/internal/components/graphs/graph-empty';
-import { getGraphVectorStats, getMetricType } from './utils';
+import { getGraphVectorStats, getMetricType, getYTickValues, sortResources } from './utils';
 
 const chartPropsValue = {
   chartHeight: 175,
@@ -27,23 +27,6 @@ const chartLegendPropsValue = {
   symbolSpacer: 7,
   height: 30,
   gutter: 10,
-};
-
-const getYTickValues = (value: number): number[] => [
-  _.round(value / 4, 1),
-  _.round(value / 2, 1),
-  _.round((3 * value) / 4, 1),
-  _.round(value, 1),
-  _.round((5 * value) / 4, 1),
-  _.round((6 * value) / 4, 1),
-];
-
-const sortResources: SortResourcesProps = (a, b) => {
-  const aVal = _.get(a, 'values');
-  const bVal = _.get(b, 'values');
-  const x = _.get(a, ['values', aVal.length - 1, 1]);
-  const y = _.get(b, ['values', bVal.length - 1, 1]);
-  return y - x;
 };
 
 const getMaxCapacity = (topConsumerStatsResult: PrometheusResponse['data']['result']) => {
@@ -122,8 +105,3 @@ type TopConsumerBodyProps = {
   metricType?: string;
   sortByOption?: string;
 };
-
-type SortResourcesProps = (
-  a: PrometheusResponse['data']['result'],
-  b: PrometheusResponse['data']['result'],
-) => number;
