+++ frontend/packages/ceph-storage-plugin/src/components/dashboard-page/storage-dashboard/top-consumers-card/utils.ts
@@ -14,8 +14,30 @@ export const getGraphVectorStats: GetStats = (response, metricType, unit) => {
   });
 };
 
+export const getYTickValues = (value: number): number[] => [
+  _.round(value / 4, 1),
+  _.round(value / 2, 1),
+  _.round((3 * value) / 4, 1),
+  _.round(value, 1),
+  _.round((5 * value) / 4, 1),
+  _.round((6 * value) / 4, 1),
+];
+
+export const sortResources: SortResourcesProps = (a, b) => {
+  const aVal = _.get(a, 'values');
+  const bVal = _.get(b, 'values');
+  const x = _.get(a, ['values', aVal.length - 1, 1]);
+  const y = _.get(b, ['values', bVal.length - 1, 1]);
+  return y - x;
+};
+
 type GetStats = (
   response: PrometheusResponse['data']['result'],
   metric?: string,
   unit?: string,
 ) => DataPoint[][];
+
+type SortResourcesProps = (
+  a: PrometheusResponse['data']['result'],
+  b: PrometheusResponse['data']['result'],
+) => number;
